class event_teams + inserevent_teams

+++ team.php
@@ -7,16 +7,9 @@
 </head>
 <body>
     <?php
-        //KONEKSI DATABASE
-        $mysqli = new mysqli("localhost", "root", "", "esport");
-        if($mysqli->connect_errno){ //kalau ada nilai selain 0, maka ada KEGAGALAN KONEKSI
-            echo "Koneksi database gagal: ".$mysqli->connect_error;
-            exit();
-        }
-        //MEMBUAT PERINTAH SELECT ALL MOVIE
-        $stmt = $mysqli->prepare("SELECT t.idteam, t.name, g.name as game FROM team t INNER JOIN game g ON t.idgame = g.idgame"); //prepare mencegah sql injection
-        $stmt->execute(); 
-        $res = $stmt->get_result(); //res = result untuk menampung value hasil
+        require_once 'classteam.php';
+        $team = new Team();
+        $res = $team->getAllTeam();
 
         //BUAT TABEL
         echo "<table border = '1'>";
@@ -30,11 +23,10 @@
                 <td>".$row['name']."</td>
                 <td>".$row['game']."</td>
                 <td><a href='editteam.php?idteam=".$row['idteam']."'>Ubah Data</a></td>
-                <td><a href='deleteteam.php?idteam=".$row['idteam']."'>Hapus Data</a></td>
+                <td><a href='deleteteam.php?idteam=".$row['idteam']."'onclick='return confirm(\"Apakah Anda yakin ingin menghapus Team ini?\");'>Hapus Data</a></td>
             </tr>";
         }
         echo "</table>";
-        $mysqli->close();
     ?>
     <a href="insertteam.php">Add New Team</a>
 </body>
